Sway Animation Complete
Applied Shaders to the foliage with an appropriate speed.

However, the scripts is completed, there, and applied to ground foliage, it is not working? So imma take the L.

But if someone wants to see the sway animation in action and lmk if the sway speed in good, that would be neat! If there is any adjusted needed, feel free to go into the WindMat and adjust the speed etc to a better liking with the sliders!